# backend/src/AosAdjutant.Api/Features/Units/UnitAttackProfileController.cs
using AosAdjutant.Api.Common;
using AosAdjutant.Api.Features.AttackProfiles;
using AosAdjutant.Api.Features.WeaponEffects;
using Microsoft.AspNetCore.Mvc;

namespace AosAdjutant.Api.Features.Units;

[Route("api/units/{unitId:int}/attack-profiles")]
[ApiController]
[Tags("Attack Profiles")]
public sealed class UnitAttackProfileController(AttackProfileService attackProfileService)
    : ControllerBase
{
    [HttpPost]
    [EndpointSummary("Create an attack profile under a unit")]
    [ProducesResponseType<AttackProfileResponseDto>(StatusCodes.Status201Created)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status404NotFound)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status409Conflict)]
    public async Task<ActionResult<AttackProfileResponseDto>> CreateAttackProfile(
        [FromRoute] int unitId,
        [FromBody] CreateAttackProfileDto attackProfileData
    )
    {
        var attackProfileResult = await attackProfileService.CreateAttackProfile(
            unitId,
            attackProfileData
        );
        return attackProfileResult.Match(
            ap =>
                CreatedAtAction(
                    nameof(AttackProfileController.GetAttackProfile),
                    "AttackProfile",
                    new { attackProfileId = ap.AttackProfileId },
                    new AttackProfileResponseDto(
                        ap.AttackProfileId,
                        ap.Name,
                        ap.IsRanged,
                        ap.Range,
                        ap.Attacks,
                        ap.ToHit,
                        ap.ToWound,
                        ap.Rend,
                        ap.Damage,
                        ap.UnitId,
                        ap.Version,
                        ap.WeaponEffects.Select(wp => new WeaponEffectResponseDto(wp.Key, wp.Name))
                            .ToList()
                    )
                ),
            this.ApiProblem
        );
    }

    [HttpGet]
    [EndpointSummary("Get all attack profiles for a unit")]
    [ProducesResponseType<PaginatedResponse<AttackProfileResponseDto>>(StatusCodes.Status200OK)]
    [ProducesResponseType<ProblemDetails>(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<PaginatedResponse<AttackProfileResponseDto>>> GetAttackProfiles(
        [FromRoute] int unitId,
        [FromQuery] AttackProfileQuery attackProfileQuery
    )
    {
        var attackProfilesResult = await attackProfileService.GetUnitAttackProfiles(
            unitId,
            attackProfileQuery
        );
        return attackProfilesResult.Match(
            attackProfiles =>
                Ok(
                    attackProfiles.Map(ap => new AttackProfileResponseDto(
                        ap.AttackProfileId,
                        ap.Name,
                        ap.IsRanged,
                        ap.Range,
                        ap.Attacks,
                        ap.ToHit,
                        ap.ToWound,
                        ap.Rend,
                        ap.Damage,
                        ap.UnitId,
                        ap.Version,
                        ap.WeaponEffects.Select(wp => new WeaponEffectResponseDto(wp.Key, wp.Name))
                            .ToList()
                    ))
                ),
            this.ApiProblem
        );
    }
}