# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AosAdjutant.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/units/{unitId}/attack-profiles": {
      "post": {
        "tags": [
          "Attack Profiles"
        ],
        "summary": "Create an attack profile under a unit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttackProfileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttackProfileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttackProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AttackProfileResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttackProfileResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttackProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Attack Profiles"
        ],
        "summary": "Get all attack profiles for a unit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IsRanged",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AttackProfileSortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAttackProfileResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAttackProfileResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAttackProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttackProfileResponseDto": {
        "required": [
          "attackProfileId",
          "name",
          "isRanged",
          "range",
          "attacks",
          "toHit",
          "toWound",
          "rend",
          "damage",
          "unitId",
          "version",
          "weaponEffects"
        ],
        "type": "object",
        "properties": {
          "attackProfileId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "isRanged": {
            "type": "boolean"
          },
          "range": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "attacks": {
            "type": "string"
          },
          "toHit": {
            "type": "integer",
            "format": "int32"
          },
          "toWound": {
            "type": "integer",
            "format": "int32"
          },
          "rend": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "damage": {
            "type": "string"
          },
          "unitId": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          },
          "weaponEffects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeaponEffectResponseDto"
            }
          }
        }
      },
      "AttackProfileSortBy": {
        "enum": [
          "Name",
          "IsRanged",
          null
        ]
      },
...
      "CreateAttackProfileDto": {
        "required": [
          "name",
          "isRanged",
          "attacks",
          "toHit",
          "toWound",
          "damage",
          "weaponEffects"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "isRanged": {
            "type": "boolean"
          },
          "range": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "attacks": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "toHit": {
            "type": "integer",
            "format": "int32"
          },
          "toWound": {
            "type": "integer",
            "format": "int32"
          },
          "rend": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "damage": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "weaponEffects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PaginatedResponseOfAttackProfileResponseDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttackProfileResponseDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ]
      },
...
      "WeaponEffectResponseDto": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}